+++ CHANGELOG.md
@@ -5,6 +5,12 @@ PrestaShop 1.6.x, 1.7.x and 8.x
 
 # Changelog
 
+## [3.8.0]
+### Added
+- Add configuration to support asynchronous processing of AltaPay payment callbacks.
+### Fixed
+- Fix: Display nonformatted amount in capture input.
+
 ## [3.7.8]
 ### Fixed
 - Fix: Full amount is being captured instead of the updated order total amount after the order has been modified.